WIP: Stop trying to upset members based on their phone number
In general since our member phone number data quality is so poor
(we have many members with duplicate phone numbers), we can't rely
on matching against the phone number when pulling Spoke data into
Identity.

When pulling messages and opt-outs, if an external id for a Spoke
Campaign contact is set, use that when both pulling messages - the
external id came from Identity so we know what member is involved.

If not we must create a new member every time - because if there is
zero matching members, we need to create one; if there is one matching
member, that member could have a bad phone number, so we can't assume
it's them, so create a new one; if there are multiple members with
matching numbers, we can't pick the right one reliably, so create a
new one.

For Spoke users, upsert based on their known email address, because
that is more reliable.

In all cases, update external ids in Spoke when one is not known, so
if multiple contacts are processed for a newly upserted member, the
same member is used for all further contacts in the same campaign.